fix(entity-storage): query layer deny-by-default — drop Neutral rows (audit C-6, #1714 follow-through)
SqlEntityQuery::filterCandidates() kept any row that was not Forbidden
(!isForbidden()), so an access-Neutral row — "no policy granted a view" —
survived the query/count layer, contradicting the entity gate's deny-by-default
(isAllowed()) that every consumer applies. #1714 wired the real composed access
handler into getQuery(); this flips the residual predicate so the query layer
enforces deny-by-default itself instead of leaning on consumers to re-deny.

- filterCandidates(): survivor test is now isAllowed(); count() inherits it.
- The empty/unwired-handler fallback now denies every candidate (fail-closed)
  instead of passing the unfiltered window through — closing the open-by-default
  primitive the audit's section G flags. Production wiring is unchanged (#1714
  accessHandlerResolver); the fallback is reached only by direct construction or
  a missing resolver.

Consumer fallout, all verified and fixed:
- GraphQL / discovery / semantic integration harnesses built SqlEntityStorage
  WITHOUT the #1714 accessHandlerResolver, so getQuery() used the empty fallback
  while the real policies lived on the endpoint guard. Post-flip the query layer
  denied rows before the guard could allow them. Wired the harness storage with
  the resolver exactly as AbstractKernel/EntityStorageFactory do — the query
  layer and the guard now share one composed handler, matching production.
- Genealogy SSR (a pre-existing production defect this surfaces): the controller
  threaded the viewer account into the relationship-topology traversal, so under
  #1714 + deny-by-default the wired relationship policy dropped every edge with a
  non-viewable endpoint — collapsing the "Private living relative" placeholders
  the SSR concealment renders. Topology is now gathered in system context
  (accessCheck(false)); concealment stays per-person via the gate (neighborSlots
  / ancestorGenerationsRedacted), so living/private relatives show as redacted
  placeholders rather than vanishing.

Failing-first: SqlEntityQueryDeniesNeutralRowTest (replaces the deliberate
SqlEntityQueryNeutralAdmitsRowTest pin) asserts Neutral rows are dropped at both
execute() and count(), Allowed rows survive, and an unwired handler fails closed;
all returned the admitted/pass-through result against the pre-flip code.
access-control.md Layer 3 corrected (was documented as an open candidate window).

Verified: Unit 8811 green, Integration 1305 green, genealogy 16 green,
check-getquery-bindings / cs-check / phpstan green.
