+++ kitty-ops/01KVQRCKSVEH9W7H20G6Y1X1N9.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQRCKSVEH9W7H20G6Y1X1N9", "profile_id": "researcher-robbie", "action": "specify", "request_text": "Execute mission step contract specify (software-dev/specify).\nStep capture_intent: Conduct or verify the human discovery interview, then capture confirmed user intent\n\nUnresolved delegation candidates: 010-specification-fidelity-requirement, 037-living-documentation-sync", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "2df2354a9e06f710", "governance_context_available": true, "started_at": "2026-06-22T13:30:19.279684+00:00"}
+{"event": "completed", "invocation_id": "01KVQRCKSVEH9W7H20G6Y1X1N9", "completed_at": "2026-06-22T13:30:19.291480+00:00", "outcome": "done", "closed_by": "agent"}
