+++ kitty-ops/01KVQRCNXM2WRYW15NT92P8S8W.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQRCNXM2WRYW15NT92P8S8W", "profile_id": "researcher-robbie", "action": "specify", "request_text": "Execute mission step contract specify (software-dev/specify).\nStep map_examples: Capture concrete rules, examples, and open questions for the behavior slice\nUnresolved delegation candidates: example-mapping-workshop", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "2df2354a9e06f710", "governance_context_available": true, "started_at": "2026-06-22T13:30:21.440149+00:00"}
+{"event": "completed", "invocation_id": "01KVQRCNXM2WRYW15NT92P8S8W", "completed_at": "2026-06-22T13:30:21.453677+00:00", "outcome": "done", "closed_by": "agent"}
