+++ kitty-ops/01KVQRCR43RTVS96EG89R5XT5P.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQRCR43RTVS96EG89R5XT5P", "profile_id": "researcher-robbie", "action": "specify", "request_text": "Execute mission step contract specify (software-dev/specify).\nStep validate_requirements: Validate requirements against spec fidelity rules\nResolved delegations: tactic:requirements-validation-workflow", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "2df2354a9e06f710", "governance_context_available": true, "started_at": "2026-06-22T13:30:21.456364+00:00"}
+{"event": "completed", "invocation_id": "01KVQRCR43RTVS96EG89R5XT5P", "completed_at": "2026-06-22T13:30:21.468224+00:00", "outcome": "done", "closed_by": "agent"}
