+++ kitty-ops/01KVQRCT5000J01Y4ZGRKS2QAW.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQRCT5000J01Y4ZGRKS2QAW", "profile_id": "researcher-robbie", "action": "specify", "request_text": "Execute mission step contract specify (software-dev/specify).\nStep write_spec: Write spec.md to kitty-specs/{feature}/\nDeclared command: Write spec.md in kitty-specs/{mission_slug}/\nCommand status: declared only; the host/operator owns execution.", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "2df2354a9e06f710", "governance_context_available": true, "started_at": "2026-06-22T13:30:25.733753+00:00"}
+{"event": "completed", "invocation_id": "01KVQRCT5000J01Y4ZGRKS2QAW", "completed_at": "2026-06-22T13:30:25.747183+00:00", "outcome": "done", "closed_by": "agent"}
