+++ kitty-ops/01KVQRDBPNW5BT10SD171ZKVMQ.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQRDBPNW5BT10SD171ZKVMQ", "profile_id": "architect-alphonso", "action": "plan", "request_text": "Execute mission step contract plan (software-dev/plan).\nStep bootstrap: Load charter context for this action\nDeclared command: spec-kitty charter context --action plan --role plan --json [--profile {wp.agent_profile}] [--tool {env.agent_tool}]\nCommand status: declared only; the host/operator owns execution.", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "aad43954bae76784", "governance_context_available": true, "started_at": "2026-06-22T13:30:44.349391+00:00"}
+{"event": "completed", "invocation_id": "01KVQRDBPNW5BT10SD171ZKVMQ", "completed_at": "2026-06-22T13:30:44.362001+00:00", "outcome": "done", "closed_by": "agent"}
