+++ kitty-ops/01KVQRDEDB1NXZENWEVWHV12DW.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQRDEDB1NXZENWEVWHV12DW", "profile_id": "architect-alphonso", "action": "plan", "request_text": "Execute mission step contract plan (software-dev/plan).\nStep read_spec: Read and understand the mission specification\nUnresolved delegation candidates: 010-specification-fidelity-requirement", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "aad43954bae76784", "governance_context_available": true, "started_at": "2026-06-22T13:30:46.495009+00:00"}
+{"event": "completed", "invocation_id": "01KVQRDEDB1NXZENWEVWHV12DW", "completed_at": "2026-06-22T13:30:46.507684+00:00", "outcome": "done", "closed_by": "agent"}
