+++ kitty-ops/01KVQRDGG7NZEVP9EM3CXPVBJK.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQRDGG7NZEVP9EM3CXPVBJK", "profile_id": "architect-alphonso", "action": "plan", "request_text": "Execute mission step contract plan (software-dev/plan).\nStep design_architecture: Design the implementation architecture and component breakdown\nResolved delegations: tactic:requirements-validation-workflow, tactic:adr-drafting-workflow", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "aad43954bae76784", "governance_context_available": true, "started_at": "2026-06-22T13:30:48.659892+00:00"}
+{"event": "completed", "invocation_id": "01KVQRDGG7NZEVP9EM3CXPVBJK", "completed_at": "2026-06-22T13:30:48.672630+00:00", "outcome": "done", "closed_by": "agent"}
