+++ kitty-ops/01KVQRDJHC1WFWJE2GH7Y4QV6Z.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQRDJHC1WFWJE2GH7Y4QV6Z", "profile_id": "architect-alphonso", "action": "plan", "request_text": "Execute mission step contract plan (software-dev/plan).\nStep write_plan: Write plan.md to kitty-specs/{feature}/\nDeclared command: Write plan.md in kitty-specs/{mission_slug}/\nCommand status: declared only; the host/operator owns execution.", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "aad43954bae76784", "governance_context_available": true, "started_at": "2026-06-22T13:30:50.763807+00:00"}
+{"event": "completed", "invocation_id": "01KVQRDJHC1WFWJE2GH7Y4QV6Z", "completed_at": "2026-06-22T13:30:50.775787+00:00", "outcome": "done", "closed_by": "agent"}
