+++ kitty-ops/01KVQRDJM14N881CZEYEGSH28N.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQRDJM14N881CZEYEGSH28N", "profile_id": "architect-alphonso", "action": "plan", "request_text": "Execute mission step contract plan (software-dev/plan).\nStep document_decisions: Record architectural decisions as ADRs\nUnresolved delegation candidates: 003-decision-documentation-requirement", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "aad43954bae76784", "governance_context_available": true, "started_at": "2026-06-22T13:30:48.570050+00:00"}
+{"event": "completed", "invocation_id": "01KVQRDJM14N881CZEYEGSH28N", "completed_at": "2026-06-22T13:30:48.582618+00:00", "outcome": "done", "closed_by": "agent"}
