+++ kitty-ops/01KVQRDMNHWRE99QZGT2RHVDNT.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQRDMNHWRE99QZGT2RHVDNT", "profile_id": "architect-alphonso", "action": "plan", "request_text": "Execute mission step contract plan (software-dev/plan).\nStep commit_plan: Commit the plan to main branch\nUnresolved delegation candidates: 029-agent-commit-signing-policy, 033-targeted-staging-policy", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "aad43954bae76784", "governance_context_available": true, "started_at": "2026-06-22T13:30:52.932511+00:00"}
+{"event": "completed", "invocation_id": "01KVQRDMNHWRE99QZGT2RHVDNT", "completed_at": "2026-06-22T13:30:52.944129+00:00", "outcome": "done", "closed_by": "agent"}
