+++ kitty-ops/01KVQRECT4H5AZJCZ9CT99T2AG.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQRECT4H5AZJCZ9CT99T2AG", "profile_id": "architect-alphonso", "action": "tasks", "request_text": "Execute mission step contract tasks (software-dev/tasks).\nStep bootstrap: Load charter context for this action\nDeclared command: spec-kitty charter context --action tasks --role tasks --json [--profile {wp.agent_profile}] [--tool {env.agent_tool}]\nCommand status: declared only; the host/operator owns execution.", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "4b1ef4fc1e04c2cf", "governance_context_available": true, "started_at": "2026-06-22T13:31:18.137165+00:00"}
+{"event": "completed", "invocation_id": "01KVQRECT4H5AZJCZ9CT99T2AG", "completed_at": "2026-06-22T13:31:18.149681+00:00", "outcome": "done", "closed_by": "agent"}
