+++ kitty-ops/01KVQREFD6HR0BQ54E8D53E1GM.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQREFD6HR0BQ54E8D53E1GM", "profile_id": "architect-alphonso", "action": "tasks", "request_text": "Execute mission step contract tasks (software-dev/tasks).\nStep outline: Produce tasks.md \u2014 work packages translated from implementation concerns in the plan\nResolved delegations: tactic:problem-decomposition, tactic:requirements-validation-workflow", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "4b1ef4fc1e04c2cf", "governance_context_available": true, "started_at": "2026-06-22T13:31:20.429907+00:00"}
+{"event": "completed", "invocation_id": "01KVQREFD6HR0BQ54E8D53E1GM", "completed_at": "2026-06-22T13:31:20.441844+00:00", "outcome": "done", "closed_by": "agent"}
