+++ kitty-ops/01KVQREHMT9Z0PC2QAQ57Z8J6G.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQREHMT9Z0PC2QAQ57Z8J6G", "profile_id": "architect-alphonso", "action": "tasks", "request_text": "Execute mission step contract tasks (software-dev/tasks).\nStep packages: Generate individual tasks/WP##.md prompt files\nUnresolved delegation candidates: 010-specification-fidelity-requirement, 024-locality-of-change", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "4b1ef4fc1e04c2cf", "governance_context_available": true, "started_at": "2026-06-22T13:31:22.464722+00:00"}
+{"event": "completed", "invocation_id": "01KVQREHMT9Z0PC2QAQ57Z8J6G", "completed_at": "2026-06-22T13:31:22.477327+00:00", "outcome": "done", "closed_by": "agent"}
