+++ kitty-ops/01KVQRNRN37SQ79PSHSSDXV0CC.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQRNRN37SQ79PSHSSDXV0CC", "profile_id": "architect-alphonso", "action": "tasks", "request_text": "Execute mission step contract tasks (software-dev/tasks).\nStep outline: Produce tasks.md \u2014 work packages translated from implementation concerns in the plan\nResolved delegations: tactic:problem-decomposition, tactic:requirements-validation-workflow", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "4b1ef4fc1e04c2cf", "governance_context_available": true, "started_at": "2026-06-22T13:35:18.952955+00:00"}
+{"event": "completed", "invocation_id": "01KVQRNRN37SQ79PSHSSDXV0CC", "completed_at": "2026-06-22T13:35:18.964734+00:00", "outcome": "done", "closed_by": "agent"}
