+++ kitty-ops/01KVQRNTJF77P4SS8Y7VGQTD2Z.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQRNTJF77P4SS8Y7VGQTD2Z", "profile_id": "architect-alphonso", "action": "tasks", "request_text": "Execute mission step contract tasks (software-dev/tasks).\nStep packages: Generate individual tasks/WP##.md prompt files\nUnresolved delegation candidates: 010-specification-fidelity-requirement, 024-locality-of-change", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "4b1ef4fc1e04c2cf", "governance_context_available": true, "started_at": "2026-06-22T13:35:20.944586+00:00"}
+{"event": "completed", "invocation_id": "01KVQRNTJF77P4SS8Y7VGQTD2Z", "completed_at": "2026-06-22T13:35:20.956297+00:00", "outcome": "done", "closed_by": "agent"}
