+++ kitty-ops/01KVQTQJ1W196CSJM93F76SAGP.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQTQJ1W196CSJM93F76SAGP", "profile_id": "researcher-robbie", "action": "specify", "request_text": "Execute mission step contract specify (software-dev/specify).\nStep bootstrap: Load charter context for this action\nDeclared command: spec-kitty charter context --action specify --role specify --json [--profile {wp.agent_profile}] [--tool {env.agent_tool}]\nCommand status: declared only; the host/operator owns execution.", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "2df2354a9e06f710", "governance_context_available": true, "started_at": "2026-06-22T14:11:15.670054+00:00"}
+{"event": "completed", "invocation_id": "01KVQTQJ1W196CSJM93F76SAGP", "completed_at": "2026-06-22T14:11:15.683319+00:00", "outcome": "done", "closed_by": "agent"}
