+++ kitty-ops/01KVQTQVRP6ZM00AMM5XS3E009.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQTQVRP6ZM00AMM5XS3E009", "profile_id": "researcher-robbie", "action": "specify", "request_text": "Execute mission step contract specify (software-dev/specify).\nStep document_decisions: Record any design decisions made during specification\nUnresolved delegation candidates: 003-decision-documentation-requirement", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "2df2354a9e06f710", "governance_context_available": true, "started_at": "2026-06-22T14:11:25.048136+00:00"}
+{"event": "completed", "invocation_id": "01KVQTQVRP6ZM00AMM5XS3E009", "completed_at": "2026-06-22T14:11:25.062753+00:00", "outcome": "done", "closed_by": "agent"}
