+++ kitty-ops/01KVQTQXSCG4RVYXCT3JRWTRN4.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQTQXSCG4RVYXCT3JRWTRN4", "profile_id": "researcher-robbie", "action": "specify", "request_text": "Execute mission step contract specify (software-dev/specify).\nStep commit_spec: Commit the specification to main branch\nUnresolved delegation candidates: 029-agent-commit-signing-policy, 033-targeted-staging-policy", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "2df2354a9e06f710", "governance_context_available": true, "started_at": "2026-06-22T14:11:27.216796+00:00"}
+{"event": "completed", "invocation_id": "01KVQTQXSCG4RVYXCT3JRWTRN4", "completed_at": "2026-06-22T14:11:27.229261+00:00", "outcome": "done", "closed_by": "agent"}
