+++ kitty-ops/01KVQTQXXK0AM1FCVKBHV3F6WH.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQTQXXK0AM1FCVKBHV3F6WH", "profile_id": "researcher-robbie", "action": "specify", "request_text": "Execute mission step contract specify (software-dev/specify).\nStep write_spec: Write spec.md to kitty-specs/{feature}/\nDeclared command: Write spec.md in kitty-specs/{mission_slug}/\nCommand status: declared only; the host/operator owns execution.", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "2df2354a9e06f710", "governance_context_available": true, "started_at": "2026-06-22T14:11:27.808899+00:00"}
+{"event": "completed", "invocation_id": "01KVQTQXXK0AM1FCVKBHV3F6WH", "completed_at": "2026-06-22T14:11:27.820787+00:00", "outcome": "done", "closed_by": "agent"}
