+++ kitty-ops/01KVQTR3F3JS40W6N72WMES3X3.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQTR3F3JS40W6N72WMES3X3", "profile_id": "architect-alphonso", "action": "plan", "request_text": "Execute mission step contract plan (software-dev/plan).\nStep bootstrap: Load charter context for this action\nDeclared command: spec-kitty charter context --action plan --role plan --json [--profile {wp.agent_profile}] [--tool {env.agent_tool}]\nCommand status: declared only; the host/operator owns execution.", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "aad43954bae76784", "governance_context_available": true, "started_at": "2026-06-22T14:11:33.483886+00:00"}
+{"event": "completed", "invocation_id": "01KVQTR3F3JS40W6N72WMES3X3", "completed_at": "2026-06-22T14:11:33.496551+00:00", "outcome": "done", "closed_by": "agent"}
