+++ kitty-ops/01KVQTR64PS3FKNT7WA8XGFQHP.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQTR64PS3FKNT7WA8XGFQHP", "profile_id": "architect-alphonso", "action": "plan", "request_text": "Execute mission step contract plan (software-dev/plan).\nStep read_spec: Read and understand the mission specification\nUnresolved delegation candidates: 010-specification-fidelity-requirement", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "aad43954bae76784", "governance_context_available": true, "started_at": "2026-06-22T14:11:35.560519+00:00"}
+{"event": "completed", "invocation_id": "01KVQTR64PS3FKNT7WA8XGFQHP", "completed_at": "2026-06-22T14:11:35.571857+00:00", "outcome": "done", "closed_by": "agent"}
