+++ kitty-ops/01KVQTR85FJ0J35FFT5YD9K07W.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQTR85FJ0J35FFT5YD9K07W", "profile_id": "architect-alphonso", "action": "plan", "request_text": "Execute mission step contract plan (software-dev/plan).\nStep design_architecture: Design the implementation architecture and component breakdown\nResolved delegations: tactic:requirements-validation-workflow, tactic:adr-drafting-workflow", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "aad43954bae76784", "governance_context_available": true, "started_at": "2026-06-22T14:11:37.678422+00:00"}
+{"event": "completed", "invocation_id": "01KVQTR85FJ0J35FFT5YD9K07W", "completed_at": "2026-06-22T14:11:37.690063+00:00", "outcome": "done", "closed_by": "agent"}
