+++ kitty-ops/01KVQTRA7ME6ZWE293AQR2DWSV.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQTRA7ME6ZWE293AQR2DWSV", "profile_id": "architect-alphonso", "action": "plan", "request_text": "Execute mission step contract plan (software-dev/plan).\nStep document_decisions: Record architectural decisions as ADRs\nUnresolved delegation candidates: 003-decision-documentation-requirement", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "aad43954bae76784", "governance_context_available": true, "started_at": "2026-06-22T14:11:39.790391+00:00"}
+{"event": "completed", "invocation_id": "01KVQTRA7ME6ZWE293AQR2DWSV", "completed_at": "2026-06-22T14:11:39.802254+00:00", "outcome": "done", "closed_by": "agent"}
