+++ kitty-ops/01KVQTRC9MAPBF35BK926NFS4N.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQTRC9MAPBF35BK926NFS4N", "profile_id": "architect-alphonso", "action": "plan", "request_text": "Execute mission step contract plan (software-dev/plan).\nStep write_plan: Write plan.md to kitty-specs/{feature}/\nDeclared command: Write plan.md in kitty-specs/{mission_slug}/\nCommand status: declared only; the host/operator owns execution.", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "aad43954bae76784", "governance_context_available": true, "started_at": "2026-06-22T14:11:41.951413+00:00"}
+{"event": "completed", "invocation_id": "01KVQTRC9MAPBF35BK926NFS4N", "completed_at": "2026-06-22T14:11:41.963134+00:00", "outcome": "done", "closed_by": "agent"}
