+++ kitty-ops/01KVQTRED5NKG8TWDYFQC62NRC.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQTRED5NKG8TWDYFQC62NRC", "profile_id": "architect-alphonso", "action": "plan", "request_text": "Execute mission step contract plan (software-dev/plan).\nStep commit_plan: Commit the plan to main branch\nUnresolved delegation candidates: 029-agent-commit-signing-policy, 033-targeted-staging-policy", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "aad43954bae76784", "governance_context_available": true, "started_at": "2026-06-22T14:11:44.084590+00:00"}
+{"event": "completed", "invocation_id": "01KVQTRED5NKG8TWDYFQC62NRC", "completed_at": "2026-06-22T14:11:44.096469+00:00", "outcome": "done", "closed_by": "agent"}
