+++ kitty-ops/01KVQTRYVTDDQMAZ009X3BA413.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQTRYVTDDQMAZ009X3BA413", "profile_id": "architect-alphonso", "action": "tasks", "request_text": "Execute mission step contract tasks (software-dev/tasks).\nStep outline: Produce tasks.md \u2014 work packages translated from implementation concerns in the plan\nResolved delegations: tactic:problem-decomposition, tactic:requirements-validation-workflow", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "4b1ef4fc1e04c2cf", "governance_context_available": true, "started_at": "2026-06-22T14:12:00.783658+00:00"}
+{"event": "completed", "invocation_id": "01KVQTRYVTDDQMAZ009X3BA413", "completed_at": "2026-06-22T14:12:00.795998+00:00", "outcome": "done", "closed_by": "agent"}
