+++ kitty-ops/01KVQTS0SPX6GYKQ1VY89GG24Y.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQTS0SPX6GYKQ1VY89GG24Y", "profile_id": "architect-alphonso", "action": "tasks", "request_text": "Execute mission step contract tasks (software-dev/tasks).\nStep packages: Generate individual tasks/WP##.md prompt files\nUnresolved delegation candidates: 010-specification-fidelity-requirement, 024-locality-of-change", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "4b1ef4fc1e04c2cf", "governance_context_available": true, "started_at": "2026-06-22T14:12:02.766098+00:00"}
+{"event": "completed", "invocation_id": "01KVQTS0SPX6GYKQ1VY89GG24Y", "completed_at": "2026-06-22T14:12:02.777604+00:00", "outcome": "done", "closed_by": "agent"}
