+++ kitty-ops/01KVQTS2QS3P9RVW27WQEJW9PG.jsonl
@@ -1,0 +1,2 @@
+{"event": "started", "invocation_id": "01KVQTS2QS3P9RVW27WQEJW9PG", "profile_id": "architect-alphonso", "action": "tasks", "request_text": "Execute mission step contract tasks (software-dev/tasks).\nStep finalize: Validate dependencies and finalize WP metadata\nDeclared command: spec-kitty agent mission finalize-tasks\nCommand status: declared only; the host/operator owns execution.\nUnresolved delegation candidates: 024-locality-of-change", "actor": "claude", "mode_of_work": "task_execution", "governance_context_hash": "4b1ef4fc1e04c2cf", "governance_context_available": true, "started_at": "2026-06-22T14:12:04.791552+00:00"}
+{"event": "completed", "invocation_id": "01KVQTS2QS3P9RVW27WQEJW9PG", "completed_at": "2026-06-22T14:12:04.806201+00:00", "outcome": "done", "closed_by": "agent"}
