+++ kitty-ops/lifecycle.jsonl
@@ -1,0 +1,24 @@
+{"agent": "claude", "at": "2026-06-22T12:12:18.701749+00:00", "canonical_action_id": "discovery::research", "mission_id": "01KVQKG48RJ9X8FGHCETS7X260", "phase": "started", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T12:17:26.879926+00:00", "canonical_action_id": "discovery::research", "mission_id": "01KVQKG48RJ9X8FGHCETS7X260", "phase": "completed", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T12:17:27.574706+00:00", "canonical_action_id": "specify::specify", "mission_id": "01KVQKG48RJ9X8FGHCETS7X260", "phase": "started", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T12:19:25.975452+00:00", "canonical_action_id": "specify::specify", "mission_id": "01KVQKG48RJ9X8FGHCETS7X260", "phase": "completed", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T12:21:17.842159+00:00", "canonical_action_id": "plan::plan", "mission_id": "01KVQKG48RJ9X8FGHCETS7X260", "phase": "started", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T12:26:32.001320+00:00", "canonical_action_id": "plan::plan", "mission_id": "01KVQKG48RJ9X8FGHCETS7X260", "phase": "completed", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T12:27:59.911831+00:00", "canonical_action_id": "tasks::tasks", "mission_id": "01KVQKG48RJ9X8FGHCETS7X260", "phase": "started", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T12:33:45.665995+00:00", "canonical_action_id": "tasks::tasks", "mission_id": "01KVQKG48RJ9X8FGHCETS7X260", "phase": "completed", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T13:29:52.842977+00:00", "canonical_action_id": "discovery::research", "mission_id": "framework-cleanup-alpha245-01KVQQMG", "phase": "started", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T13:30:03.582922+00:00", "canonical_action_id": "discovery::research", "mission_id": "framework-cleanup-alpha245-01KVQQMG", "phase": "completed", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T13:30:04.197412+00:00", "canonical_action_id": "specify::specify", "mission_id": "framework-cleanup-alpha245-01KVQQMG", "phase": "started", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T13:30:13.207893+00:00", "canonical_action_id": "specify::specify", "mission_id": "framework-cleanup-alpha245-01KVQQMG", "phase": "completed", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T13:30:28.369770+00:00", "canonical_action_id": "plan::plan", "mission_id": "framework-cleanup-alpha245-01KVQQMG", "phase": "started", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T13:30:40.474856+00:00", "canonical_action_id": "plan::plan", "mission_id": "framework-cleanup-alpha245-01KVQQMG", "phase": "completed", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T13:30:53.428341+00:00", "canonical_action_id": "tasks::tasks", "mission_id": "framework-cleanup-alpha245-01KVQQMG", "phase": "started", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T13:31:16.624522+00:00", "canonical_action_id": "tasks::tasks", "mission_id": "framework-cleanup-alpha245-01KVQQMG", "phase": "completed", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T14:10:47.860747+00:00", "canonical_action_id": "discovery::research", "mission_id": "framework-cleanup-alpha245-01KVQSN8", "phase": "started", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T14:10:58.917410+00:00", "canonical_action_id": "discovery::research", "mission_id": "framework-cleanup-alpha245-01KVQSN8", "phase": "completed", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T14:11:00.096697+00:00", "canonical_action_id": "specify::specify", "mission_id": "framework-cleanup-alpha245-01KVQSN8", "phase": "started", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T14:11:11.789403+00:00", "canonical_action_id": "specify::specify", "mission_id": "framework-cleanup-alpha245-01KVQSN8", "phase": "completed", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T14:11:27.739021+00:00", "canonical_action_id": "plan::plan", "mission_id": "framework-cleanup-alpha245-01KVQSN8", "phase": "started", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T14:11:29.595173+00:00", "canonical_action_id": "plan::plan", "mission_id": "framework-cleanup-alpha245-01KVQSN8", "phase": "completed", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T14:11:44.564457+00:00", "canonical_action_id": "tasks::tasks", "mission_id": "framework-cleanup-alpha245-01KVQSN8", "phase": "started", "reason": null, "wp_id": null}
+{"agent": "claude", "at": "2026-06-22T14:11:55.164591+00:00", "canonical_action_id": "tasks::tasks", "mission_id": "framework-cleanup-alpha245-01KVQSN8", "phase": "completed", "reason": null, "wp_id": null}
