+++ kitty-ops/ops-index.jsonl
@@ -1,0 +1,55 @@
+{"invocation_id": "01KVQMAXJRV7CMS206ANNRM7ER", "profile_id": "researcher-robbie", "started_at": "2026-06-22T12:19:30.065672+00:00"}
+{"invocation_id": "01KVQMBDG8AJMM3HMJTJSGNJ98", "profile_id": "researcher-robbie", "started_at": "2026-06-22T12:19:45.857197+00:00"}
+{"invocation_id": "01KVQMBXNWJP78Z1T5WWRRRH5N", "profile_id": "researcher-robbie", "started_at": "2026-06-22T12:20:02.538729+00:00"}
+{"invocation_id": "01KVQMCBNKQ9M3CZD29HESCPW1", "profile_id": "researcher-robbie", "started_at": "2026-06-22T12:20:16.564567+00:00"}
+{"invocation_id": "01KVQMCWGRS7RSC3JSE8GD4QKQ", "profile_id": "researcher-robbie", "started_at": "2026-06-22T12:20:34.080092+00:00"}
+{"invocation_id": "01KVQMDBNRR13P51G48SYR9TMB", "profile_id": "researcher-robbie", "started_at": "2026-06-22T12:20:49.553973+00:00"}
+{"invocation_id": "01KVQMDW4RATXMJ9WMT6AETAQD", "profile_id": "researcher-robbie", "started_at": "2026-06-22T12:21:06.285060+00:00"}
+{"invocation_id": "01KVQMQXE9GHXZXA43T9N3T4T9", "profile_id": "architect-alphonso", "started_at": "2026-06-22T12:26:35.849642+00:00"}
+{"invocation_id": "01KVQMRAR7ANQ9W52VSK5ECZFE", "profile_id": "architect-alphonso", "started_at": "2026-06-22T12:26:48.814029+00:00"}
+{"invocation_id": "01KVQMRSNN8HC8WDZMYBGKE927", "profile_id": "architect-alphonso", "started_at": "2026-06-22T12:27:04.165113+00:00"}
+{"invocation_id": "01KVQMS6B18NS905QATC7XVJ6R", "profile_id": "architect-alphonso", "started_at": "2026-06-22T12:27:17.117792+00:00"}
+{"invocation_id": "01KVQMSNFZP0VGTY55S4PWH3QY", "profile_id": "architect-alphonso", "started_at": "2026-06-22T12:27:32.610491+00:00"}
+{"invocation_id": "01KVQMT24R53QC41W9GHN8YFV1", "profile_id": "architect-alphonso", "started_at": "2026-06-22T12:27:45.611801+00:00"}
+{"invocation_id": "01KVQN54ZJRETKPKC1RD5J9ZMA", "profile_id": "architect-alphonso", "started_at": "2026-06-22T12:33:49.372574+00:00"}
+{"invocation_id": "01KVQN5JXD36ZBN51BECRX13F2", "profile_id": "architect-alphonso", "started_at": "2026-06-22T12:34:03.071168+00:00"}
+{"invocation_id": "01KVQN61WMWM6884B34CQKKXBY", "profile_id": "architect-alphonso", "started_at": "2026-06-22T12:34:18.386647+00:00"}
+{"invocation_id": "01KVQN6FCDD2G3BHZJHJWK6D8Q", "profile_id": "architect-alphonso", "started_at": "2026-06-22T12:34:32.301354+00:00"}
+{"invocation_id": "01KVQRCH4B8SNZMHABH30M06DW", "profile_id": "researcher-robbie", "started_at": "2026-06-22T13:30:17.103830+00:00"}
+{"invocation_id": "01KVQRCKSVEH9W7H20G6Y1X1N9", "profile_id": "researcher-robbie", "started_at": "2026-06-22T13:30:19.279684+00:00"}
+{"invocation_id": "01KVQRCNXM2WRYW15NT92P8S8W", "profile_id": "researcher-robbie", "started_at": "2026-06-22T13:30:21.440149+00:00"}
+{"invocation_id": "01KVQRCR43RTVS96EG89R5XT5P", "profile_id": "researcher-robbie", "started_at": "2026-06-22T13:30:21.456364+00:00"}
+{"invocation_id": "01KVQRCR1QV94FCKAGGH5FMJW2", "profile_id": "researcher-robbie", "started_at": "2026-06-22T13:30:23.600823+00:00"}
+{"invocation_id": "01KVQRCT5000J01Y4ZGRKS2QAW", "profile_id": "researcher-robbie", "started_at": "2026-06-22T13:30:25.733753+00:00"}
+{"invocation_id": "01KVQRCW7HZYTJ8VNMMJPRAA9T", "profile_id": "researcher-robbie", "started_at": "2026-06-22T13:30:27.886279+00:00"}
+{"invocation_id": "01KVQRDBPNW5BT10SD171ZKVMQ", "profile_id": "architect-alphonso", "started_at": "2026-06-22T13:30:44.349391+00:00"}
+{"invocation_id": "01KVQRDEDB1NXZENWEVWHV12DW", "profile_id": "architect-alphonso", "started_at": "2026-06-22T13:30:46.495009+00:00"}
+{"invocation_id": "01KVQRDGG7NZEVP9EM3CXPVBJK", "profile_id": "architect-alphonso", "started_at": "2026-06-22T13:30:48.659892+00:00"}
+{"invocation_id": "01KVQRDJM14N881CZEYEGSH28N", "profile_id": "architect-alphonso", "started_at": "2026-06-22T13:30:48.570050+00:00"}
+{"invocation_id": "01KVQRDJHC1WFWJE2GH7Y4QV6Z", "profile_id": "architect-alphonso", "started_at": "2026-06-22T13:30:50.763807+00:00"}
+{"invocation_id": "01KVQRDMNHWRE99QZGT2RHVDNT", "profile_id": "architect-alphonso", "started_at": "2026-06-22T13:30:52.932511+00:00"}
+{"invocation_id": "01KVQRECT4H5AZJCZ9CT99T2AG", "profile_id": "architect-alphonso", "started_at": "2026-06-22T13:31:18.137165+00:00"}
+{"invocation_id": "01KVQREFD6HR0BQ54E8D53E1GM", "profile_id": "architect-alphonso", "started_at": "2026-06-22T13:31:20.429907+00:00"}
+{"invocation_id": "01KVQREHMT9Z0PC2QAQ57Z8J6G", "profile_id": "architect-alphonso", "started_at": "2026-06-22T13:31:22.464722+00:00"}
+{"invocation_id": "01KVQREKMK42TY1QRAYDB43SRB", "profile_id": "architect-alphonso", "started_at": "2026-06-22T13:31:24.594638+00:00"}
+{"invocation_id": "01KVQRNP40AAERX3KWR3JFVKW9", "profile_id": "architect-alphonso", "started_at": "2026-06-22T13:35:16.984744+00:00"}
+{"invocation_id": "01KVQRNRN37SQ79PSHSSDXV0CC", "profile_id": "architect-alphonso", "started_at": "2026-06-22T13:35:18.952955+00:00"}
+{"invocation_id": "01KVQRNTJF77P4SS8Y7VGQTD2Z", "profile_id": "architect-alphonso", "started_at": "2026-06-22T13:35:20.944586+00:00"}
+{"invocation_id": "01KVQRNWHBW28AYNDYCQX609Q9", "profile_id": "architect-alphonso", "started_at": "2026-06-22T13:35:23.152867+00:00"}
+{"invocation_id": "01KVQTQJ1W196CSJM93F76SAGP", "profile_id": "researcher-robbie", "started_at": "2026-06-22T14:11:15.670054+00:00"}
+{"invocation_id": "01KVQTQMR7RB3CACCXNV90XFX1", "profile_id": "researcher-robbie", "started_at": "2026-06-22T14:11:17.861205+00:00"}
+{"invocation_id": "01KVQTQPWG2NWXZJ6YCVBV567B", "profile_id": "researcher-robbie", "started_at": "2026-06-22T14:11:20.037572+00:00"}
+{"invocation_id": "01KVQTQS29S0B37N4VZN5NTTC5", "profile_id": "researcher-robbie", "started_at": "2026-06-22T14:11:22.863251+00:00"}
+{"invocation_id": "01KVQTQVRP6ZM00AMM5XS3E009", "profile_id": "researcher-robbie", "started_at": "2026-06-22T14:11:25.048136+00:00"}
+{"invocation_id": "01KVQTQXXK0AM1FCVKBHV3F6WH", "profile_id": "researcher-robbie", "started_at": "2026-06-22T14:11:27.808899+00:00"}
+{"invocation_id": "01KVQTQXSCG4RVYXCT3JRWTRN4", "profile_id": "researcher-robbie", "started_at": "2026-06-22T14:11:27.216796+00:00"}
+{"invocation_id": "01KVQTR3F3JS40W6N72WMES3X3", "profile_id": "architect-alphonso", "started_at": "2026-06-22T14:11:33.483886+00:00"}
+{"invocation_id": "01KVQTR64PS3FKNT7WA8XGFQHP", "profile_id": "architect-alphonso", "started_at": "2026-06-22T14:11:35.560519+00:00"}
+{"invocation_id": "01KVQTR85FJ0J35FFT5YD9K07W", "profile_id": "architect-alphonso", "started_at": "2026-06-22T14:11:37.678422+00:00"}
+{"invocation_id": "01KVQTRA7ME6ZWE293AQR2DWSV", "profile_id": "architect-alphonso", "started_at": "2026-06-22T14:11:39.790391+00:00"}
+{"invocation_id": "01KVQTRC9MAPBF35BK926NFS4N", "profile_id": "architect-alphonso", "started_at": "2026-06-22T14:11:41.951413+00:00"}
+{"invocation_id": "01KVQTRED5NKG8TWDYFQC62NRC", "profile_id": "architect-alphonso", "started_at": "2026-06-22T14:11:44.084590+00:00"}
+{"invocation_id": "01KVQTRWD2W14FBDJTRGSPA1NY", "profile_id": "architect-alphonso", "started_at": "2026-06-22T14:11:58.796621+00:00"}
+{"invocation_id": "01KVQTRYVTDDQMAZ009X3BA413", "profile_id": "architect-alphonso", "started_at": "2026-06-22T14:12:00.783658+00:00"}
+{"invocation_id": "01KVQTS0SPX6GYKQ1VY89GG24Y", "profile_id": "architect-alphonso", "started_at": "2026-06-22T14:12:02.766098+00:00"}
+{"invocation_id": "01KVQTS2QS3P9RVW27WQEJW9PG", "profile_id": "architect-alphonso", "started_at": "2026-06-22T14:12:04.791552+00:00"}
